backend/requirements.txt
fastapi
pandas
sqlalchemy
mysql-connector-python
psycopg2-binary
psycopg_binary
pgvector
langchain_postgres
uvicorn
pydantic
pydantic[email]
passlib
pyjwt
python-dotenv
pytest
langchain
langgraph
langchain-core
langchain-classic
langchain-openai
langchain-groq
langchain-ollama
langchain-community
autopep8
prettier
coloredlogs
langchain-huggingface
pypdf
python-multipart
alembic
sentence-transformers
rank_bm25
duckdb
slowapi
fastapi-mail